+++ src/tokens/typography.js
@@ -1,0 +1,10 @@
+export const sizeXXL = '20px';
+export const sizeXL = '18px';
+export const sizeL = '15px';
+export const sizeM = '13px';
+export const sizeS = '12px';
+export const sizeXS = '11px';
+
+export const bold = 'bold';
+export const semibold = 'semibold';
+export const regular = 'regular';